+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Search: Search by name/names to allow regex match.
 - Studio: Enhance search to be by name, regex, or attribute query.
 - Studio: Refactor actions to implement base class.
+- Studio: Add help screen.
 - Test: Test for studio.
 ### Fixed:
 - Query: Query method for inequality to cover cases where attribute is not present.